requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-login>=0.6
flask-wtf>=1.2
flask-limiter>=3.5
requests>=2.31
python-dotenv>=1.0
pyotp>=2.9
qrcode>=7.4
pillow>=10.0
apscheduler>=3.10
scikit-learn>=1.4
pandas>=2.2
gunicorn>=22.0
pytest>=8.0
